+++ packages/blog/tsconfig.app.json
@@ -33,7 +33,8 @@
       "@graphql/*": ["./src/graphql/*"],
       "@pages/*": ["./src/pages/*"],
       "@recoil/*": ["./src/recoil/*"],
-      "@hooks/*": ["./src/hooks/*"]
+      "@hooks/*": ["./src/hooks/*"],
+      "@utils/*": ["./src/utils/*"]
     }
   },
   "include": ["src"]
